[IROptimizer] Reuse copy propagation check and fix a bug
Refactor the code to have one shared way of checking if we are in a copy
propagation case. This also fixes a bug where we were not checking that two
intervals had the same values and thus wrongly assuming this was a copy
propagation.

The added test case demonstrates the issue and would fail without the change in
the IR optimizer.