Perform Variable CSE on Variables of small enough size
For Variables that are 256 elements or smaller and of the same type, are private, and have no train kind, this will attempt to combine them into a single shared variable. This is particularly useful for IntLookupTables of the same "kind" (e.g. Tanh lookup tables), which often use the same data as their lookup table Var.