Add TextTranslator program for driving text translation
*Description*:
This commit adds a program for driving text translation, similar to the ImageClassifier for image classification. It does pre and post processing of the data, converting between input text and dictionary tokens, as well as determining the highest likelihood sentence from the beam search output of the model.

It is loosely based on loading a model from pytorch_translate (https://github.com/pytorch/translate/tree/master/pytorch_translate/cpp). It includes command-line parameters for setting the model's max input and output sentence lengths, as well as beam size. It also allows for specifying a length penalty when determining the highest likelihood output sentence (this is for post-processing the output translation, separate from the model design).

*Testing*: It has been tested with unrolled en2gr protos. Shareable protos will be coming soon and uploaded to aws.

*Documentation*: Added info on running to `Testing.md`