Introduce a NodeHandle type to be used for referencing Nodes by other Nodes
Until now, NodeValue was used to refer to the results of Nodes. But unfortunately, every NodeValue would add itself to the use-list of a Node it references, which is rather expensive. In many cases, NodeValue objects are created as temporary objects and it makes no sense to consider them to be uses.

We want to count only references from Nodes to the results of other Nodes as uses. To achieve that this commit introduces a new type NodeHandle. It will be used instead of NodeValue inside Nodes to refer to the results of other nodes. All NodeHandles add themselves as uses to the use lists. And the former NodeValue will be changed in such a way that it does not register itself in the use-lists anymore.