[GraphIR] Add a proper copy constructor for NodeHandle
Prior to this patch, copying a NodeHandle around (like with a vector::resize)
would not properly set the tracking of the value.

For instance, this was impossible to pretty print nodes with vector of inputs
like concat, because our pretty printer was doing creating a local NodeHandle.

This patch features a proper copy constructor for NodeHandle and at the same
time, make sure pretty printing vectors of NodeHandle doesn't create temporary
objects.

NFC.