[graph-optimizer] Avoid useless calls of the simplifyNode transformation
The call of simplifyNode for the RHS was ignored when the call of simplifyNode for the LHS was successful. It was ignored even if RHS was actually simplified and thus any nodes created by this simplification were produced, but never used and later removed by the dead code elimination pass.