Update the gather node to support Batches.
Update the gather node to support Batches. In this commit we add the
'batches' operand that marks some indices in the beginning of the tensor
as 'batch' indices. The gather operation is executed on each sample in
the batch independently. The batches are later concatenated. This allows
us to implement the Gather and BatchedGather operation in a single node.