[Graph IR & IR] Remove the overwritten attribute on QuantizationProfile
The QuantizationProfile node suffers the same problem as the save
node: it has implicit memory dependencies and the scheduler could
be tricker to generate invalid code because of that.

Unlike the save node, the problem is less likely to happen because
this node only deals with its only created variables. Yet, as soon
as those variables are created, it is possible to temper with them
to create invalid schedules.

Unlike the save node, where we taught the scheduler how to deal
with the implicit memory dependencies, this patch fixes the
representation of the QuantizationProfile node (and its underlying
QuantizationProfileInstr) by removing the implicit memory
dependencies. After this patch, the save node is the only node
with implicit memory dependencies.

Test Plan:
$ ninja check

$ ./bin/image-classifier tests/images/imagenet/*.png -image_mode=0to1 -m=resnet50 -dump_profile=resnet50_prof_async.yaml

$ ./bin/image-classifier tests/images/imagenet/*.png -image_mode=0to1 -m=resnet50 -dump_profile=resnet50_prof_async2.yaml

$ diff -q resnet50_prof_async2.yaml resnet50_prof_async.yaml

$ ./bin/image-classifier tests/images/imagenet/*.png -image_mode=0to1 -m=resnet50 -dump_profile=resnet50_prof_async2.yaml

$ diff -q resnet50_prof_async2.yaml resnet50_prof_async.yaml

Results:
Everything passed, no diffs.