Placeholder: add the graph creators and a simple unit test that checks that we can construct the graph.
This commit is the first few steps in the implementation of the
Placeholder node. The plan is discussed in #1334. Placeholder are
unbound variables where the content of the tensor is provided by the
execution engine or the AOT compiler at runtime.

This commit introduces Storage as a superclass for Variable (bound) and
Placeholder (unbound). This PR also adds a simple unit test that checks
that we can create new Placeholder nodes. This PR does not implement any
of the infrastructure that's required to do anything useful with
Placeholder variables, such as loading content into them, deleting them,
pretty dotty-printer, etc.

Next steps (not necessarily in any specific order):

1. Teach the execution engine to bind tensors to the Placeholder nodes.

2. Verify that dotty printing, dump() and debugging work well.

3. Cleanup the APIs that are related to Variable and Placeholder and
make them consistent.

4. Change (some of) the unit tests to use the new Placeholder API.

5. Make sure that our optimizations are correct when placeholder are
used.