[NodeValue] Provide methods to access the users of a NodeValue
Prior to this patch the only way to get to the users of a node value
was to get through the users of the related node and filtered out
the uses of that did not match the result number.

This feature is now directly accessible through the NodeValueIterator
class and exposed via the new methods NodeValue::getUsers,
NodeValue::getNumUsers, and NodeValue::hasOneUse.

This fixes #1346.