[Tensor] Fix a leak with the reset method
When using the reset method on an unowned tensor, we did not reset
this state when allocating memory. In other words, although we were
specifically allocating a new buffer, we were not taking its onwership
and were leaking it when the tensor got destroyed.

The added test case would fail leak sanitizer without that patch.