[LLVMIRGen] Make sure path delimiters don't result in empty entry names
Prior to this patch setting a name ending with '/' would produce an
empty entry name. Combined with the bundle saver, it was possible to
end up with an empty name being appended to the files saved.

E.g.,
$ model-runner -model model -emit-bundle output
Would produce in output:
model.o
model.weights

But
$ model-runner -model model/ -emit-bundle output
Would produce in output:
.o
.weights

Thus, under unix those are hidden files and one could think the bundler
did not work.

This patch makes sure the entry name is never empty by walking back the '/'
until it finds a non-empty name or the beginning of the string.