Make sure that MemoryAllocator uses wide-enough unsigned types for representing the address space being allocated
The types should be able to represent any address in the address space of the device, whose memory is being allocated by means of the MemoryAllocator. This device may be different from the host running Glow. E.g. Glow may run on a host with a 32-bit address space, but use a  MemoryAllocator for a GPU or an accelerator with a 64-bit address space.

Therefore, to be on the safe side use `uint64_t` instead of `size_t` in MemoryAllocator APIs.

Update the user's of the API to use `uint64_t` instead of `size_t`.