GLOW_ASSERT and GLOW_UNREACHABLE should write messages to stderr
As GLOW_ASSERT and GLOW_UNREACHABLE are supposed to be the drop-in replacements for `assert` and `llvm_unreachable`, they should use the same output stream, which is stderr. It also helps with testing, as some testing frameworks like `gtest` allow for testing the assertions and expect the messages to be written to stderr.