[Importer][Caffe2] Fix the loading of batched MatMul with transpose
For batchedMatMul LHS, (Transpose RHS), we used to hardcode the transpose
indices to {1, 0}. RHS could actually have more than two dimensions and the
Caffe2 specs says all the dimensions are untouched but the last two are
swapped.

The added test case doesn't demonstrate that specific problem because more
than two dimensions are not supported for the loading of the batched matmul
proper. Instead, this adds a test case for the supported broadcast of RHS.