[GraphOptimizer] Fix/simplify optimizeArithmeticNodes()
The previous version of optimizeArithmeticNodes() and simplifyNode(), which used a combination of a worklist and recursion, made this error prone, and hard to debug/understand. I simplified it by moving to just a worklist and using GraphPreOrderVisitor to ensure operands are processed before nodes themselves.

The previous version also had a bug. There were extra adds and maxes in Resnet50 because simplifyNodes() canonicalizes by creating a new node, but the old node was not being replaced by the new node. I think it was assumed it would be done recursively up the stack, but it was not the case. The added test case shows displays such a pattern. The adds and maxes were essentially being split into n new nodes given n users.

These extra nodes would have been recombined during CSE, however the optimizeArithmeticNodes() pass comes after CSE. Even still, it would be silly to rely on CSE to fix problems this pass creates.