Move ProtobufLoader output map to Variables instead of SaveNodes
There's no reason to save outputs by SaveNodes instead of Variables. It is potentially dangerous because non-quantized graphs are often erased after they are quantized, and so any references to the SaveNodes from the non-quantized graph are no longer valid.

This is the case since https://github.com/pytorch/glow/pull/1458