[Backend] Add a network name argument to the save method
This patch allows to specify the name when saving a function. This name is
intended to be prepended to any files that would be generated by the save method
and it s also used as the name of the entry point in the network.

In the context of the BundleSaver, this name will be prepended to the .o and
.weights files that are generated and will be used as the function name for
the entry point that wraps the network.

Moreover this patch decouples the naming of the input directory of the model
and the name of the generated files of the bundle.
Previously, we were using the last directory name of the input path as the
name of the output network. One of the problem with that was that our handling
of the path was bogus and we could end up with empty network names. See #1475.