[Loader] Add an option to specify the name of the network being bundled
With the previous refactoring of the save method in the backends,
the name of the network was hardcoded to be "net".

With the option -network-name, it is now possible to set the name
to whatever makes sense.

By default the network name will be derived from the last directory
name of the `-model` path.