[CPU] Improve the 2D transpose implementation.
This commit "tiles" the 2d transpose method. Processing the input matrix
in small blocks ensures that the source and destination will stay in
cache. Notice that scanning 1 row at a time and storing the scalar into
some column brings and invalidates many cache lines. The tiled scan
helps to reduce the number of used destination cache lines. I measured a
2x speedup on a small local benchmark. We could further improve this
procedure using vector instructions for the transpose of the single tile
and with recursive blocking of the 2d matrix.