Revert: [Placeholder] Add support for Placeholders in the execution engine #1586
This commit reverts the parts of the PR #1586 that pass the placeholder
variables in the run command of the execution engine. At first I thought
that this would allow to eliminate the variables from the high-level
parts of the compilation, but I realized that we need to start with
fixing all of the backends and add support to the compiled function.