Extract the variable updating parts from run().
Extract the variable updating parts from run(). This is a step in the
direction of untangling the execution engine. In this commit we remove
the variable updating logic from the run method. The next step would be
to untangle runBatch, and finally, we'll move these variable updating
functions to a new module.