[Placeholder] Implement the execution context.
The Context provides a mapping between some graph nodes, which are a
symbolic representation of some computation, and concrete tensors that
represent the inputs and outputs to the graph. The context owns the
tensors and the graph uses these values as runtime. This is useful for
the multi-threaded execution of code, where each thread has a different
execution context. The difference between this class and a regular map
is that the Context owns the Tensors (not only the pointers) and manages
their lifetime.

The plan is to pass the context to the different graph builders (for
example, the LSTM builder) and have the builder initialize the weights
inside the context. Later, the context can be used for execution by the
unit tests. We would need to split the graph builders to APIs that
require context (for example, helper builders that create new variables)
and to low-level builders that don't. It should be easy to do.

Once the Context is in place we'll be able to get rid of the last
mutable variables.