Allow for backend to accept a Function instead of IRFunction
This PR allows for Backends to take a Function pointer for `compile()` and `save()`, instead of a IRFunction. The EE now does this. This allows for example for a backend to decide whether it wants to use our Instruction level IR or implement its own. The Interpreter, CPUBackend, and OpenCL backend all still use our Instruction level IR.

Note that I needed to keep the `compile()` that takes the IRFunction around, as it is used during unit tests. I couldn't think of good ways to change the unit tests to be able to remove it. Open to suggestions.