[Placeholder] Allow differentiation of Placeholder variables.
Allow differentiation of Placeholder variables. This commit fixes a few
bugs in the scheduler, cloner and differentiation function, where we
assumed that the inputs to the graph are variables, and not
placeholders. This commit adds a unit test that checks that we can
differentiate (and later compile) functions that differentiate (and
update) placeholders.