[Context] Add a method to allocate a new backing tensor.
While writing a unit test that replaces some Variable with a Placeholder
I realized that in most cases we mainly care about allocating new
tensors, and not moving an existing tensor into the context. To
implement the new 'allocate' method I had to move the Context into the
Graph library, because it depends on the Placeholder class to figure out
the type of the Placeholder for Tensor allocatio method I had to move
the Context into the Graph library, because it depends on the
Placeholder class to figure out the type of the Placeholder for Tensor
allocation.