[GraphScheduler] Fix underflow bug
Description:
This commit fixes an underflow bug in
ChildMemSizeBasedScheduler::orderChildNodesAndSchedule that can
cause suboptimal scheduling. There is code in this function that
orders the children of a node by the amount of memory they free up
by executing. For each node, this is computed as maxSize - resultSize,
where maxSize is the total memory needed to compute and store the node's
inputs, and resultSize is the total memory needed to store the node's
outputs. For some nodes, such as a Concat that takes Variables as input,
resultSize > maxSize, so this subtraction underflows, and the scheduler
thinks that the node frees up a lot of memory and schedules it before
some or all of its siblings that actually free up more memory.

This commit fixes this bug by modifying the scheduler to use int64_t
instead of size_t to store resultSize and maxSize for all nodes.

Testing:
This commit includes a unit test that exposes this bug. This unit
tests constructs a graph in which a dummy node has two inputs:
a Concat node connected to Variables, and a Slice node connected to
a Transpose node. The Slice node should be scheduled before the
Concat node since it frees up memory, but without this patch, it is
scheduled after.

The remaining unit tests all pass.