[WIP][GraphOptimizer] Sink Transpose below Slice
Description:
This commit modifies the sinkCode function in GraphOptimizer.cpp
to sink TransposeNodes below SliceNodes. In the best case, when
all of the children of a particular TransposeNode are SliceNodes,
this optimisation will help reduce the memory footprint by taking
slices directly out of the input to the original TransposeNode
and then transposing those slices (presumably) in-place. In the
worst case, when only one of the children of a TransposeNode is
a SliceNode and the rest are nodes that the optimizer does not
sink TransposeNodes into, this optimisation will keep the memory
footprint the same because the TransposeNode that it inserts
can presumably still be executed in-place.

Testing:
This commit includes a unit test (sinkTransposeBelowSlice)
in graphOptzTest.cpp that tests this optimisation. All other unit
tests still pass.