[Graph] Add new node builders that generate Placeholder and not variables.
Add new node builders that generate Placeholders and not variables. The
new APIs will allow us to migrate our unit tests from using mutable
bound variables to unbounded placeholders. As soon as we finish the
migration we'll merge the new APIs with the regular builder methods
above. The methods are not annotated with oxygen comments because it's
only a copy that needs to go away as soon as the migration is complete.