[Quantization] Add createMapping()
Quantized `SigmoidNode` and `TanhNode` are implemented with a pre-calculated mapping that is input to a `IntLookupTable`. However implementing a quantized `LogNode` with a `IntLookupTable` requires calculating a new mapping depending on the input and output min/max. I added a helper `createMapping()` for creating such a mapping to Quantization/Base