[Support] Add a new type to represent half precision values
*Description*
The new type performs all the operation in single precision floating
point, while converting back and forth into a 16-bit storage.
The conversions are performed by the same FP16 library used by Caffe2.

*Testing*
Added specific unittests to check the basic functionality of the new
type.

*Documentation*
Mention the additional submodule dependency for the FP16 conversions.

This is related to #1329