[Tensor] Add a copyWithCast method
*Description*
When initializing tensors of different types but with semantically
the same value, it is convenient to have a method to copy the content
of one tensor to another one. This is the intent of that method.

Specifically, the use case in mind is the validation of float16 vs.
float. In that scenario, we want to create float16 tensors and use
them to initialize float tensors, then run the same program, hence
modulo the type difference, and check that the result are similar.

*Testing*
Added test case for the new method.