[Caffe2] Implement ReplaceNaN operator
Description:
This commit adds support for the ReplaceNaN Caffe2 operator to the
Caffe2 model importer. This operator accepts one tensor as an input and
a value as an argument, and replaces all elements of the input that are
NaN with the given value. To implement this operator, this commit adds a new
Glow node, IsNaN, which accepts an input tensor and outputs a tensor in
which each element is a boolean indicating whether or not the
corresponding element in the input is NaN. With this new node in
hand, a ReplaceNaN operator is implemented as a Select node with the
ReplaceNaN input as one input, a Splat node created using the replacement
value as the other input, and the output of an IsNaN node as the comparator
input. This IsNaN node becomes a new instruction called elementisnan,
which is implemented using std::isnan from the cmath header and isnan
from the math.h header in the interpreter and CPU backends,
respectively.

Testing:
This commit adds a new test case to caffe2ImporterTest.cpp which imports
and runs a Caffe2 model consisting of only the ReplaceNaN operator. This
test performs high level checks on the Glow node graph produced by the
importer, and checks that all instances of NaN in a test input tensor
are replaced with the replacement value provided in the model protobuf
and non-NaN elements have been left unchanged.

This commit also adds a new test case to OperatorTest.cpp to test that
NaN values are replaced correctly by the interpreter and CPU backends.

All other unit tests still pass.