[Node] Add a setType method
* Description *
Prior to this patch the only way to change the type of a Node was
to create a new one with a different type and manually set the
arguments to match the node you want to replace.

This patch introduces a setType method that allows to change the
type of a specific result.

This method only changes the type of the specified result, in other
words if this type is only valid if the type of the inputs are changed
as well, these inputs have to be updated independently.

E.g.,
tensor<float> A
tensor<float> B
tensor<float> = batchAdd A, B

Changing the type of the result of batchAdd won't change the type
of A and B. Thus if those are not updated the verifier will complain
that the type for the sources and the result are incompatible.

* Testing *
Added test case.

* Documentation *
N/A

Related to #1329, we want an easy way to convert graphs to different
types for testing purposes.