[Graph][Utils] Add a mutate result type function
* Description *
Add a function to mutate the element type of all the results of
a node matching another type element type.

This is going to be useful to do conversion from say FloatTy to
Float16Ty at the function scale.

* Testing *
Added test case

Related to #1329, this lays the ground work to do the conversion
in the loader from float to float16.