[Tensor] Add a convertToType method
*Description*
This patch adds a convertToType method that allows to convert the content
of a tensor from its current type to a new type.

The shape of the tensor remains unchanged, but the element type
gets updated to the new type.

For now, this method only supports conversion between float and
float16_t.

*Testing*
Added test case.

Related to #1329, this patch adds the functionality required to
change the content of a network while doing conversion from one
type to another.