[Caffe2] Implement DotProduct operator
Description:
This commit fixes #1701 and adds support for the `DotProduct` Caffe2
operator to the Caffe2 model importer. This operator accepts two tensors
of the same shape and with at most 2 dimensions, and computes their
row-wise dot product. This dot product operator is immediately lowered
to a `Mul` node if the inputs are 1D, and to a `Mul` node followed by a
`BatchedReduceAdd` on `axis=1` if the inputs are 2D.

Testing:
This commit adds two test cases to `caffe2ImporterTest.cpp`, one for 1D
inputs and the other for 2D inputs.

All other unit tests pass.