[Graph] Add a function to mutate one type to another for all the nodes of a function
*Desription*
This patch adds a function to mutate one type to another for a given function.
The methods updated the type of all the nodes matching the desired type,
including the backing tensors for the storage nodes.

This implements converting one graph from one type (e.g., float) to
another (e.g., float16_t).

*Testing*
Added test cases. In particular, we test both that the graph is
changed the way we expect and that it runs correctly.

Related to #1329.