[Graph] Make Module non-copyable
*Description*
Copying Modules around would make them share the same underlying memory
for all the functions, variables, and placeholders.

Therefore, if we were to copy a module by mistake (e.g., by forgetting
to put & after the typename of a local variable), when this copy
would get deleted, it would leave the original module with dangling
pointers.

This patch also adds a clear method for the one test case where the
copy of a module was made on purpose.

*Testing*
Added test case for the clear mehod and changed the one test case
that was using Module copy.