[Loader] Add an option to convert fp32 model to fp16
*Description*
This patch adds an option `-use-fp16` to convert all fp32 operators
into fp16 ones.

We may want to expose more conversion options in the future as well
as choosing which nodes are okay to convert, but for testing purposes
we support only one mode and a all of nothing approach.

*Testing*
Able to convert resnet50, but cannot run yet as the interpreter still
lacks some fp16 operators.

*Documentation*
None so far, we want the option to be functional on resnet50 before
documenting it.

Related to #1329