[Optimizer] Teach the bacthNorm(Conv) optimization about FP16
*Description*
Prior to this patch the optimization that merges convolution followed
by batchNorm would assert on fp16 because the handles expect a regular
single precision floating point type.

This patch templatizes this part of the optimization so that it
works seamlessly with single or half precision floating point numbers.

*Testing*
Added test case.

Related to #1329