[Graph] Allow float16 batch normalization to be created
*Description*
Prior to this patch creating a batch normalization node would
create a bunch of fp32 variables with it. When the input type
is fp16 we want these variables to be fp16 as well.

This patch fixes that by using the type of the input to create
the variables.

*Testing*
New test case added