CPU: change the way that we generate the JIT runtime
Use a custom linker language to change the linker invocation for the CPU
Runtime.  This allows us to properly track the dependencies and ensure that
Ninja does not rebuild the target unnecessarily.  As a small benefit, the linker
invocation is now also more portable.  The compiler rule still prevents the use
of the Makefile generator :-(.