[unittests][GraphScheduler] Avoid the use of a temporary std::vector
*Description*
We use a temporary std::vector because we compare directly two iterator
to check which instruction comes first in the final schedule.
Instead if we compare the distance from the beginning for each
iterator, we don't need our iterator to support random accesses and
we can avoid the temporary object.

This is what this patch does.

NFC