[Placeholder] Remove the ability to create SaveNode that saves into a variable.
This commit removes the SaveNode constructor that saves into a variable.
It also fixes the lowering method and makes it create a placeholder for
the GSUM storage.  It also renames createSaveNodePH into createSaveNode.

At the moment we need to initialize the context after lowering because
lowering creates new placeholders. When we migrate the context into the
execution method we'll be able to move the initialization outside of the
compilation code.