Generate OpenCL kernel inc header files from original OpenCL kernel sources directly using the new include-bin utility
It is now possible to treat the OpenCL sources as real C source files in the editors, formatting tools, etc.

It should also solve issues with compilers which could not handle raw string literals that are too long.

Also take the opportunity to run format.sh on the OpenCL sources we have.