[Quantization] Specialize the FunctionConverter to quantize a function
*Description*
This patch refactors the code of the function that quantizes a function
to reuse the framework set by the FunctionConverter.

This patch is not strictly speaking NFC because unlike the original
implementation the conversion process does not leave the original,
not use after conversion, node around. Other than that, it should
be NFC.

*Testing*
Ran the quantizationTest
Ran different quantized profiles (asymmetric, symmetric, symmetric
with uint8) with and without the patch for resnet50 and checked
that the output IR were identical (modulo some variable names).

Related to #1329, we provide a unified way for converting function
that fp16 is going to use as discussed in #1747.