[Graph&IR] Add a ConvertTo node
*Description*
This node/instruction represents a type conversion (a.k.a. a cast)
from the input type to the output type.
The shape of the input and output type must be identical, only the
element type is allowed to change.

This node will allow us to represent type conversions when converting
a graph from one type to another.

This patch adds the support for this node in the interpreter and
teaches the differentiation mechanism how to deal with ConvertTo
node.

Note: We could technically introduce different nodes for the different
conversion (e.g., fpext, fptrunc, inttofp, and so on) since this
node would have different interpretation based on the input and output
types. Right now, we decided not to do that because we expect this
node to not hit any backends. In particular, this node is only
accessible through the C++ API.

*Testing*
Added tests.