[Converter] Add a helper class to converter a function from type A to type B
*Description*
This patch specializes the FunctionConverter class to feature a converter
from one element type to another.
In essence this new TypeAToTypeBFunctionConverter class scans the
function, morph the node from TypeA to TypeB and inserts the right
conversions to/from typeA and typeB around the converted nodes.

This initial version follows the same schema as the quantization
of function, meaning there are no attempt to eliminate useless conversion.
This feature will be added later.

*Testing*
Tests added.