[Function] Fix a nasty bug in eraseNode(Node *)
*Description*
Prior to this patch, Function::eraseNode(Node *N) would erase a node
with the same value as N, but not necessarily N. This means that
as soon as two nodes were identical, there was a non-zero chance
that deleting one would actually make the other one being a dangling
pointer.

*Testing*
Added test case.
The added test case would assert (Could not find node to delete) in
debug mode and segmentation fault in release mode without this patch.