[Converter] Optimize intermediate conversions away
*Description*
After converting a model we often end up with patterns like this:
conversion (conversion A to B) to C
Where A and C are actually the same type.

This patch adds a method to optimize away this kind of conversions.

This optimization is not performed during the main conversion procedure
(FunctionConverter::convert) because it potentially changes the
semantic of the program when B would have incured some precision loss
(e.g., float(int64_t(float (A))) != float(A)).

This works seamlessly for the FunctionQuantizer and the
TypeAToTypeBFunctionConverter.

*Testing*
Added test case.