[GraphOptimizer] Optimize intermediate conversions away
*Description*
After converting a model we often end up with patterns like this:
conversion (conversion A to B) to C
Where A and C are actually the same type.

This patch adds a method to optimize away this kind of conversions
as well as the conversions applied to constants.

*Testing*
Added test cases.