[Verifier] Check that the type of a constant and its tensor matches
*Description*
Right now, the type of a constant and the type of its underlying tensor
are stored in two different places. As a result, it is possible to
modify one without modifying the other ending up in a situation
where the constant and its tensor have different types.

This patch makes sure that the verifier reports this problem.

*Testing*
Added test case